--- src/main/java/guiTable/controllers/AvailableBoatsController.java
@@ -1,8 +1,0 @@
-package guiTable.controllers;
-
-public class AvailableBoatsController {
-
-    public void init() {
-
-    }
-}
