--- src/main/java/guiTable/controllers/BoardController.java
@@ -1,8 +1,0 @@
-package guiTable.controllers;
-
-public class BoardController {
-    
-    public void init() {
-
-    }
-}
