+++ src/main/resources/fxml/IhmTable/ClassicPlacementPhase.fxml
@@ -1,5 +1,10 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
+<?import java.lang.*?>
+<?import javafx.scene.control.*?>
+<?import javafx.scene.layout.*?>
+<?import javafx.scene.shape.*?>
+<?import javafx.scene.text.*?>
 <?import javafx.scene.control.Button?>
 <?import javafx.scene.control.Label?>
 <?import javafx.scene.control.Menu?>
@@ -13,7 +18,7 @@
 <?import javafx.scene.shape.Rectangle?>
 <?import javafx.scene.text.Font?>
 
-<AnchorPane id="AnchorPane" fx:id="anchorPane" prefHeight="600.0" prefWidth="800.0" xmlns="http://javafx.com/javafx/8.0.141" xmlns:fx="http://javafx.com/fxml/1" fx:controller="guiTable.controller.ClassicPlacementPhaseController">
+<AnchorPane id="AnchorPane" fx:id="anchorPane" prefHeight="600.0" prefWidth="800.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="guiTable.controllers.ClassicPlacementPhaseController">
    <children>
       <MenuBar prefHeight="25.0" prefWidth="812.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="0.0">
         <menus>
@@ -214,14 +219,10 @@
             <Rectangle fx:id="sousMarinRectangle" arcHeight="5.0" arcWidth="5.0" fill="#ababab" height="29.0" layoutX="560.0" layoutY="142.0" stroke="BLACK" strokeType="INSIDE" width="99.0" />
             <Rectangle fx:id="torpilleurRectangle" arcHeight="5.0" arcWidth="5.0" fill="#ababab" height="29.0" layoutX="560.0" layoutY="176.0" stroke="BLACK" strokeType="INSIDE" width="64.0" />            
             <Rectangle fx:id="porteAvionsRectangle" arcHeight="5.0" arcWidth="5.0" fill="#ababab" height="29.0" layoutX="560.0" layoutY="40.0" stroke="BLACK" strokeType="INSIDE" width="169.0" />
-            <Button  fx:id="valider" layoutX="450.0" layoutY="505.0" mnemonicParsing="false" text="Valider" />
+            <Button fx:id="valider" layoutX="450.0" layoutY="505.0" mnemonicParsing="false" text="Valider" />
             <Button layoutX="14.0" layoutY="505.0" mnemonicParsing="false" text="Quitter" />
          </children>
       </Pane>
-      <Button layoutX="550.0" layoutY="571.0" mnemonicParsing="false" prefHeight="30.0" prefWidth="250.0" text="Envoyer" AnchorPane.bottomAnchor="0.0" AnchorPane.rightAnchor="0.0">
-         <font>
-            <Font size="14.0" />
-         </font>
-      </Button>
+      <AnchorPane fx:id="chatPane" layoutX="548.0" layoutY="274.0" prefHeight="324.0" prefWidth="250.0" />
    </children>
 </AnchorPane>
